Update CAPI File Names, Adjust Formatting to Match Current Standard, Adjust #include Names, Change #pragma once to #ifndef/#define format
